Moved DiveGear CfgWeapons to its own file
